+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.6.4](https://github.com/in2workspace/in2-wallet-ui/releases/tag/v1.6.4)
+### Changed
+- Unsigned credentials are now automaticly updated if issuer has signed them
+- Changed signature button for an info button when credential is unsigned
 
 ## [v1.6.3](https://github.com/in2workspace/in2-wallet-ui/releases/tag/v1.6.3)
 ### Fixed